pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whisperbox"
version = "1.0.2"
description = "WhisperBox - Record and transcribe audio with ease"
authors = [
    {name = "Ty Fiero", email = "tyfierodev@gmail.com"},
    {name = "Mike Bird", email = "tooluseai@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "pydub>=0.25.1",
    "pyaudio>=0.2.13",
    "rich>=13.3.0",
    "pyyaml>=6.0",
    "anthropic>=0.43.0",
    "openai>=1.59.7",
    "ollama>=0.4.6",
    "groq>=0.15.0",
    "toga>=0.4.8",
    "toga-cocoa>=0.4.8",
    "inquirerpy>=0.3.4",
    "requests>=2.32.3",
    "pyperclip>=1.9.0",
    "sounddevice>=0.5.1",
]

[project.scripts]
wb = "whisperbox.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/whisperbox"]
include = [
    "src/whisperbox/profiles/*.yaml",
    "src/whisperbox/scripts/*.py"
]